# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Online Learning System API",
    "description": "Backend API for Online Learning Platform - Challenges 1-3\n\n## Challenges Completed:\n- **Challenge 1**: Database Design (PostgreSQL, 7 tables, 3NF normalized)\n- **Challenge 2**: CRUD Operations (Courses & Classes)\n- **Challenge 3**: Authentication (Register, Login, JWT, Profile)\n\n## Features:\n- ✅ User Registration & Login with JWT\n- ✅ Role-based Authorization (admin, instructor, student)\n- ✅ Course Management (CRUD with instructor/admin)\n- ✅ Class Management (CRUD with pagination)\n- ✅ Pagination & Filtering\n- ✅ Error Handling & Validation\n- ✅ Logging & Monitoring\n\n## Test Credentials:\n- **Email**: student1@example.com\n- **Password**: password123\n- **Role**: student",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/api/classes": {
      "get": {
        "tags": ["Classes"],
        "summary": "List Classes",
        "operationId": "listClasses",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "courseId",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Classes retrieved"
          }
        }
      },
      "post": {
        "tags": ["Classes"],
        "summary": "Create Class",
        "operationId": "createClass",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["course_id", "class_name", "start_date", "end_date"],
                "properties": {
                  "course_id": {
                    "type": "integer"
                  },
                  "class_name": {
                    "type": "string"
                  },
                  "start_date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "end_date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "max_students": {
                    "type": "integer",
                    "default": 30
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Class created"
          }
        }
      }
    },
    "/api/classes/{id}": {
      "get": {
        "tags": ["Classes"],
        "summary": "Get Class by ID",
        "operationId": "getClassById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Class retrieved"
          }
        }
      },
      "put": {
        "tags": ["Classes"],
        "summary": "Update Class",
        "operationId": "updateClass",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "class_name": {
                    "type": "string"
                  },
                  "start_date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "end_date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "max_students": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Class updated"
          }
        }
      },
      "delete": {
        "tags": ["Classes"],
        "summary": "Delete Class",
        "operationId": "deleteClass",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Class deleted"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token received from /api/auth/login endpoint"
      }
    }
  }
}